Split cluster.go into focused files; delete dead code
cluster.go had grown to 398 lines mixing the HDBSCAN HTTP client, the
Run pipeline, the List summary, the per-path representative picker,
helpers, and two large unused legacy functions (BuildClusters +
cosineSim + the exported Cluster type — replaced by Run in task 2.5,
never called externally, safe to delete).

New layout:
- cluster.go (43) — package doc + Config / Result / ClusterSummary
  types. The package's public contract.
- client.go (56) — HdbscanClient and its sidecar request shape.
- representative.go (42) — representative struct + pickRepresentative.
- pipeline.go (174) — Run() decomposed into collectRepresentatives,
  reduceTo2D, runHDBSCAN, writeBackClusterIDs, summarize, and
  writeCoordsTSV. Each helper is independently readable and testable.
- list.go (54) — List() for read-only cluster summaries.

representative_test.go (renamed from cluster_test.go) unchanged —
still the only file that covers a testable unit here.

Net: -99 lines, single-responsibility per file, no file over 175.
