+++ search/cluster/representative_test.go
